🚚 Move Rendering & Scenes
- Created a scene class that application can use.
- Moved testing code for textures and rendering to a new TestScene.
- Cleanup Application code and comment.
- Added Transformation to shader.

+++ Terra/Source/Core/DeltaTime.cpp
@@ -27,8 +27,8 @@ namespace Terra
         m_Accumulator -= m_FrameTime;
     }
 
-    double DeltaTime::Get() const
+    float DeltaTime::Get() const
     {
-        return m_FrameTime;
+        return static_cast<float>(m_FrameTime);
     }
 }
